x/protocols/types/message_create_protocol_test.go
package types

import (
	"testing"

	"github.com/bitbadges/bitbadgeschain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateProtocol_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateProtocol
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateProtocol{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateProtocol{
				Creator: sample.AccAddress(),
				Name:    "non-empty",
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
